ai_echo_website/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
yt_dlp
requests
openai
whisper
elevenlabs
sentence-transformers
scikit-learn
joblib
numpy
